OpenGL: move buffer to its own file
